# Address performance issue caused by unnecessary invokeStaticMissingProperty calls