--- a/src/main/groovy/com/netflix/gradle/plugins/deb/DebFileVisitorStrategy.groovy
+++ b/src/main/groovy/com/netflix/gradle/plugins/deb/DebFileVisitorStrategy.groovy
@@ -7,7 +7,7 @@ import org.vafer.jdeb.producers.DataProducerLink
 
 import java.nio.file.Path
 
-import static com.netflix.gradle.plugins.utils.FileCopyDetailsUtils.getRootPath
+import static com.netflix.gradle.plugins.utils.GradleUtils.getRootPath
 
 class DebFileVisitorStrategy {
     protected final List<DataProducer> dataProducers