… src/main/groovy/com/netflix/gradle/plugins/utils/FileCopyDetailsUtils.groovy