Noticed that `buildDeb` seemed unreasonably slow in a project I was working on. A spot of profiling indicates that the problem is Groovy trying to treat a static call to a method in a parent class as a property, and burning a bunch of unnecessary time:

    groovy.lang.MetaClassImpl.invokeStaticMissingProperty(Object, String, Object, boolean) MetaClassImpl.java 89009ms
        groovy.lang.MetaClassImpl.getProperty(Class, Object, String, boolean, boolean) MetaClassImpl.java:1854
            groovy.lang.MetaClassImpl.getProperty(Class, Object, String, boolean, boolean) MetaClassImpl.java:1830
                groovy.lang.MetaClassImpl.invokeStaticMethod(Object, String, Object[]) MetaClassImpl.java:1447
                    org.codehaus.groovy.runtime.callsite.StaticMetaClassSite.callStatic(Class, Object[]) StaticMetaClassSite.java:65
                        org.codehaus.groovy.runtime.callsite.AbstractCallSite.callStatic(Class, Object, Object) AbstractCallSite.java:214
                            com.netflix.gradle.plugins.deb.DebCopyAction.visitFile(FileCopyDetailsInternal, Object) DebCopyAction.groovy:100 24825ms
                            com.netflix.gradle.plugins.deb.DebCopyAction.visitFile(FileCopyDetailsInternal, Object) DebCopyAction.groovy:101 21504ms
                            com.netflix.gradle.plugins.deb.DebCopyAction.visitFile(FileCopyDetailsInternal, Object) DebCopyAction.groovy:102 20944ms
                            com.netflix.gradle.plugins.deb.DebCopyAction.visitFile(FileCopyDetailsInternal, Object) DebCopyAction.groovy:103 19602ms
                            com.netflix.gradle.plugins.deb.DebCopyAction.visitDir(FileCopyDetailsInternal, Object) DebCopyAction.groovy:112 2131ms